Calculate weekly averages and find high frequency travel days
Tasks for this git commit session, for this coursework, were to:

- Calculate weekly averages of the following:
    * Travel distance for those traveling.
    * The weekly number of people of those staying at home.
- Identify high frequency travel days for people who traveled:
    * 10-25 miles.
    * 50-100 miles.
- Create visualisations for both prior bullet points.
    * A histogram for weekly averages.
    * A line chart for high frequency travel days.

I'll be frank, I don't fully understand the need to visualise the
weekly averages using a histogram as the result didn't seem like you
could extract much insight from it. That's all though. You can see it
under the visualisations folder.